Force (int) for the request_var here
Love,
ToonArmy

+++ titania/contributions/support.php
@@ -22,7 +22,7 @@ titania::add_lang('posting');
 $post_id = request_var('p', 0);
 $topic_id = request_var('t', 0);
 $start = request_var('start', 0);
-$limit = request_var('limit', phpbb::$config['posts_per_page']);
+$limit = request_var('limit', (int) phpbb::$config['posts_per_page']);
 
 // Load the topic and contrib items
 if ($post_id)
@@ -176,7 +176,7 @@ switch ($action)
 		else
 		{
 			$start = request_var('start', 0);
-			$limit = request_var('limit', phpbb::$config['topics_per_page']);
+			$limit = request_var('limit', (int) phpbb::$config['topics_per_page']);
 
 			titania_display_forums('contrib', titania::$contrib, false, array('start' => $start, 'limit' => $limit));
